AuthN: rename package to sync
