+++ pkg/services/authn/authnimpl/sync/orgsync_test.go
@@ -1,4 +1,4 @@
-package usersync
+package sync
 
 import (
 	"context"
